+++ packages/cubejs-client-core/index.d.ts
@@ -25,15 +25,21 @@ declare module '@cubejs-client/core' {
     method?: 'GET' | 'PUT' | 'POST' | 'PATCH'
   };
 
-  export interface ITransport {
-    request(method: string, params: Record<string, unknown>): () => Promise<void>;
+  export interface ITransportResponse<R> {
+    subscribe: <CBResult>(cb: (result: R, resubscribe: (() => Promise<CBResult>)) => CBResult) => Promise<CBResult>;
+    // Optional, supported in WebSocketTransport
+    unsubscribe?: () => Promise<void>;
+  }
+
+  export interface ITransport<R> {
+    request(method: string, params: Record<string, unknown>): ITransportResponse<R>;
   }
 
   /**
    * Default transport implementation.
    * @order 3
    */
-  export class HttpTransport implements ITransport {
+  export class HttpTransport implements ITransport<ResultSet> {
     /**
      * @hidden
      */
@@ -54,8 +60,10 @@ declare module '@cubejs-client/core' {
      * @hidden
      */
     protected credentials: TransportOptions['credentials'];
+
     constructor(options: TransportOptions);
-    request(method: string, params: any): () => Promise<any>;
+
+    public request(method: string, params: any): ITransportResponse<ResultSet>;
   }
 
   export type CubeJSApiOptions = {
@@ -66,7 +74,7 @@ declare module '@cubejs-client/core' {
     /**
      * Transport implementation to use. [HttpTransport](#http-transport) will be used by default.
      */
-    transport?: ITransport;
+    transport?: ITransport<any>;
     headers?: Record<string, string>;
     pollInterval?: number;
     credentials?: 'omit' | 'same-origin' | 'include';