+++ packages/cubejs-playground/charts-gen/react/index.js
@@ -6,11 +6,13 @@ const {
 } = require('@cubejs-templates/core');
 const { pascalCase } = require('change-case');
 const path = require('path');
+const {
+  DependencyTree,
+  AppContainer,
+  DevPackageFetcher,
+  executeCommand
+} = require('@cubejs-backend/templates');
 
-const DependencyTree = require('../dev/DependencyTree');
-const AppContainer = require('../dev/AppContainer');
-const DevPackageFetcher = require('../dev/DevPackageFetcher');
-const { executeCommand } = require('../dev/utils');
 const { generateCodeChunks } = require('./code-chunks-gen');
 const { REPOSITORY } = require('../env');