+++ packages/cubejs-playground/src/atoms/Button.tsx
@@ -92,4 +92,4 @@ StyledButton.Group = styled(Button.Group)`
   }
 `;
 
-export default StyledButton;
+export { StyledButton as Button };