+++ packages/cubejs-playground/src/atoms/index.tsx
@@ -1,3 +1,6 @@
-export { default as Alert } from './Alert';
-export { default as CubeLoader } from './CubeLoader';
-export { default as FatalError } from './FatalError';
+export * from './Alert';
+export * from './CubeLoader';
+export * from './FatalError';
+export * from './Button';
+export * from './Card';
+export * from './Popover';