+++ rust/cubestore/src/queryplanner/mod.rs
@@ -5,6 +5,7 @@ mod planning;
 pub mod pretty_printers;
 pub mod query_executor;
 pub mod serialized_plan;
+mod topk;
 pub mod udfs;
 
 use crate::config::injection::DIService;